+++ opensyde_tool/src/can_monitor/project/C_CamProHandler.cpp
@@ -1206,7 +1206,7 @@ void C_CamProHandler::m_CalcHash(uint32_t & oru32_HashValue) const
    // settings
    // CAN DLL configuration
    stw::scl::C_SclChecksums::CalcCRC32(this->mc_CustomCanDllPath.toStdString().c_str(),
-                                       this->mc_CustomCanDllPath.size(), oru32_HashValue);
+                                       static_cast<uint32_t>(this->mc_CustomCanDllPath.size()), oru32_HashValue);
    stw::scl::C_SclChecksums::CalcCRC32(&this->me_CanDllType, sizeof(this->me_CanDllType), oru32_HashValue);
 
    // filters
