+++ opensyde_tool/src/project_gui/system_views/C_PuiSvHandlerFiler.cpp
@@ -154,7 +154,7 @@ int32_t C_PuiSvHandlerFiler::h_SaveViews(const std::vector<C_PuiSvData> & orc_Vi
    int32_t s32_Retval = C_NO_ERR;
 
    orc_XmlParser.CreateAndSelectNodeChild("opensyde-system-views");
-   orc_XmlParser.SetAttributeUint32("length", orc_Views.size());
+   orc_XmlParser.SetAttributeUint32("length", static_cast<uint32_t>(orc_Views.size()));
    for (uint32_t u32_ItView = 0; (u32_ItView < orc_Views.size()) && (s32_Retval == C_NO_ERR); ++u32_ItView)
    {
       orc_XmlParser.CreateAndSelectNodeChild("opensyde-system-view");
