+++ opensyde_tool/src/project_gui/system_views/dashboard/C_PuiSvDbDataElementDisplayFormatter.cpp
@@ -59,7 +59,7 @@ void C_PuiSvDbDataElementDisplayFormatter::CalcHash(uint32_t & oru32_HashValue) 
 {
    stw::scl::C_SclChecksums::CalcCRC32(&this->q_IsActive, sizeof(this->q_IsActive), oru32_HashValue);
    stw::scl::C_SclChecksums::CalcCRC32(this->c_FormatterString.toStdString().c_str(),
-                                       this->c_FormatterString.length(), oru32_HashValue);
+                                       static_cast<uint32_t>(this->c_FormatterString.length()), oru32_HashValue);
 }
 
 //----------------------------------------------------------------------------------------------------------------------
